Add a timeout.
Opening a WebSocket can take a long time, if the network or the remote server
is down. This commit adds a 2 second timeout so the app can let the user know
that it's taking longer than expected. The library will continue to try
the connection. It's just now it informs the user more frequently.