Bug fix: Update request map before getting first reply.
This is the same bug we experience each time we write a client
library. I thought I had `acnetd` handle this, and it does. However,
when talking to local clients, the ACK for the request is immediately
followed by the reply. The WebSocket delivers them back to back so my
task waiting for the ACK `Future` to complete doesn't run until after
the reply is received.  The result is the request hangs and never gets
a reply.

The fix was to add a Future to the ACK information. The WebSocket
subscription can be paused. Both `.requestReplyStream()` and
`.requestReply()` methods pause the stream until they can set up the
lookup table.